Clone notes to the same folder
Currently notes are cloned to the default folder if the clone
note link is clicked from "All Notes" or "Starred." They are only
cloned to the same folder as the original note if the clone note
link is clicked inside of that folder.

This commit changes this behavior such that a clone is always in
the same folder as the note it was cloned from.